+++ js/task/task.component.js
@@ -225,7 +225,10 @@ angular.module('cycloneApp')
                     ctrl.currentTask.newEntryText = '';
                     ctrl.currentTask.newEntryProject = '';
                     ctrl.currentTask.newEntryTask = '';
-                    ctrl.currentTask.newEntryManualTime = '';
+                    ctrl.newEntryManualTime = '';
+
+                    // TODO: time is not yet part of the saving process
+                    // ctrl.currentTask.newEntryManualTime = '';
                     ctrl.currentTask.newEntryType = 'work';
                     ctrl.currentTask.$save().then(function(ref) {
                     }, function(error) {