+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.54 - 04.09.2017 - Refinements
 * Animation improvements
+* Save current entry (but not the manual time field yet)
 
 ## 0.53 - 01.09.2017 - Disabled auto routing of today
 * Disabled auto routing, because it made issues